refactor(app): applies suggestions from PR comments
