+++ src/components/TableContent/index.js
@@ -1,0 +1,2 @@
+import TableContent from "./TableContent";
+export default TableContent;
