+++ src/hooks/useFetchBaseCurrencyRates.js
@@ -4,18 +4,20 @@ import { fetchWithErrorHandling } from "../utils/fetchWithErrorHandling";
 import { mapDataToBaseCurrencyRates } from "../utils/mapDataToBaseCurrencyRates";
 
 export function useFetchBaseCurrencyRates(baseCurrency) {
-  const [rates, setRates] = useState([]);
+  const [baseCurrencyRates, setBaseCurrencyRates] = useState([]);
   const [error, setError] = useState(null);
   const [loading, setLoading] = useState(false);
 
   useEffect(() => {
     const fetchRates = async () => {
       setLoading(true);
-      const [data, error] = await fetchWithErrorHandling(
-        async () => await getBaseCurrencyRates(baseCurrency)
+      const [data, error] = await fetchWithErrorHandling(() =>
+        getBaseCurrencyRates(baseCurrency)
       );
-      const rates = mapDataToBaseCurrencyRates(data);
-      setRates(rates);
+      if (!error) {
+        const baseCurrencyRates = mapDataToBaseCurrencyRates(data);
+        setBaseCurrencyRates(baseCurrencyRates);
+      }
       setError(error);
       setLoading(false);
     };
@@ -23,7 +25,7 @@ export function useFetchBaseCurrencyRates(baseCurrency) {
   }, [baseCurrency]);
 
   return {
-    rates,
+    baseCurrencyRates,
     loading,
     error,
   };
