+++ src/pages/CurrenciesList.jsx
@@ -1,16 +1,13 @@
 import { useState } from "react";
-import { Link } from "react-router-dom";
 import { routes } from "../constants/routes";
 import Breadcrumbs from "../components/Breadcrumbs";
 import Heading from "../components/Heading";
 import Select from "../components/Select";
-import Table, { Thead, Tbody, Tr, Td } from "../components/Table";
-import { useFetchBaseCurrencyRates } from "../hooks/useFetchBaseCurrencyRates";
+import Table, { Thead, Tr, Tbody, Td } from "../components/Table";
+import TableContent from "../components/TableContent";
 
 function CurrenciesListPage({ currencies }) {
   const [baseCurrency, setBaseCurrency] = useState("usd");
-  const { rates, loading, error } = useFetchBaseCurrencyRates(baseCurrency);
-
   const handleSelectChange = (event) => setBaseCurrency(event.target.value);
 
   return (
@@ -36,49 +33,9 @@ function CurrenciesListPage({ currencies }) {
             <Td variant="header">Change</Td>
           </Tr>
         </Thead>
-        {loading ? (
-          error ? (
-            <Tbody>
-              <Tr>
-                <Td variant="body" isInfo>
-                  {error.message}
-                </Td>
-              </Tr>
-            </Tbody>
-          ) : (
-            <Tbody>
-              <Tr>
-                <Td variant="body" isInfo>
-                  loading...
-                </Td>
-              </Tr>
-            </Tbody>
-          )
-        ) : (
-          <Tbody>
-            {rates.map(({ name, value, change }) => {
-              const secondaryCurrency = name.split("-")[1];
-              return (
-                <Tr key={name}>
-                  <Td variant="body">
-                    <div className="currency-comparison-cell-inside-wrapper">
-                      <div>{name}</div>
-                      <Link
-                        to={`/details/${baseCurrency}/${secondaryCurrency}`}
-                      >
-                        <div className="details-vector"></div>
-                      </Link>
-                    </div>
-                  </Td>
-                  <Td variant="body" isMiddle>
-                    {value}
-                  </Td>
-                  <Td variant="body">{change}</Td>
-                </Tr>
-              );
-            })}
-          </Tbody>
-        )}
+        <Tbody>
+          <TableContent baseCurrency={baseCurrency} />
+        </Tbody>
       </Table>
     </>
   );
